+++ 01-sample/provider.tf
@@ -1,11 +1,11 @@
 provider "azurerm" {
-  version = "~>1.29.0"
+  version = "~>1.30.1"
 }
 
 provider "http" {
-  version = "~>1.1.0"
+  version = "~>1.1.1"
 }
 
 provider "random" {
-  version = "~>2.1.0"
+  version = "~>2.1.2"
 }